Initialize repository
The repository has been initialized with our default developer tooling
and an empty Rust crate. Checks have been implemented using Earthly so
that they can easily be run locally and in CI, including formatters and
linters for common file types. The project is developed as open source
under either the Apache or MIT license, as it is common in the Rust
community.