Create command-line interface for clawless
clawless will have its own command-line interface to enable users to
generate new projects and commands. This both provides a better
development experience to users and allows us to dogfood the clawless
framework as we are building it.

A minimal command-line application has been created as a starting point
to work on the framework and its features.