Collect commands using a build script
The goal with clawless is to hide most of the complexity of building a
command-line interface and allow users to focus on their commands. A
first idea that we're exploring is to use a build script to collect the
user's commands at build time and inject them into an application that
is generated using proc-macros.

This first spike adds a proof-of-concept for a build script that
collects a flat list of commands, converts them into an `enum`, and
includes that in the application.