Check GitHub Actions using zizmor
zizmor[^1] has been added as a check to the continuous integration setup
of this project. It is run both as a pre-commit hook and a job on GitHub
Actions to ensure we are following the best practices for GitHub Actions
and avoid potentially dangerous configurations.

zizmor warns that the use of `${{ matrix.target }}` statements enables
possible template injections, since it cannot statically check the
inputs to the statement. Since they are coming from the `earthly ls`
command and are further filtered by `jq`, we deem the risk of template
injection to be very low.