Prepare repository for automated releases (#11)
The repository has been prepared for automating releases to crates.io.
New labels have been added to the repository that can be used to tag
pull requests so that they are included in the changelog, and a new
GitHub Actions workflow has been added that runs after a new GitHub
release is published and pushes the crates in the repository to
crates.io. Together, these features make it easy to publish new versions
using GitHub's tooling.