Document CLIs using doc comments (#27)
The derive macros for Clawless have been expanded to capture the doc
comments on commands and use them to document the command-line
interface. The first line of the doc comment is used as a command's
short description, while the whole comment block is available as the
command's help text.

Take the following example:

```rust
/// Create a new project
///
/// This command creates a new project and sets it up for clawless.
pub async fn new(_args: NewArgs) {}
```

The first line of the doc comment is used as the command's description,
which gets printed when calling the CLI without arguments or when
passing in `-h`.

The full block is printed when the user asks for help, i.e. by typing in
`new --help`.