Return result type from commands
The interface of commands has been changed to return a `Result` so that
error handling in command-line applications can be centralized. Since
errors in commands are unrecoverable, the result type wraps the unit
type and an `anyhow::Error`.

Returning a `Result` from commands makes error handling within a command
easier and is a first step towards better support for error handling in
Clawless.