Rename the error types
The error types have been renamed to avoid confusion and avoid potential
naming conflicts in the future. The previously re-exported `Result` type
was unintuitive, since it did not follow the established convention of
including the return type with the type (e.g. `Result<()>`). By renaming
the re-export to `CommandResult`, we hope that this confusion is
avoided.

We also renamed the re-export of `anyhow::Context`, since we already
have an issue in the backlog to provide context to commands (see #32).
By renaming the context trait now, we avoid a naming conflict in the
future.