Install rustup with Flox (#98)
We are using rustup to change which Rust version we use for some of our
checks, but have not added it to the Flox environment. Since rustup
ships with the common components, we had to remove them from our Flox
manifest.

We also had to bump the MSRV and fix a lint, since our checks now seem
to work properly.