Ignore directories when linting Markdown files (#104)
Linting Markdown files failed when `cargo doc` had been run, since files
in `target/doc` did not match the standards in our project. To fix this,
we are now passing an argument to the `markdownlint-cli` that ignores
all files and directories from the `.gitignore` file. This should reduce
the amount of configuration we need to maintain and provide a good
enough filter to avoid future false negatives.