Return `Result` from `clawless::main!` (#105)
The proc macro for `clawless::main!` has been modified to return a
`Result`, making it possible to use the `?` operator when initializing
the async runtime and scheduling the user's function on it. While this
is functionally very similar to what was there before, it allows us to
remove two `unwrap` statements from our code.

This feature required bumping the minimum version of `anyhow` to fix
issues with the conversation between the different error types.