Add flag to force checks to run (#106)
The checks for latest and minimum dependency versions mess with the
local filesystem, which is why they have a guard clause that checks
whether the git repository is clean before running. This works well for
regular check runs, but is annoying when trying to fix broken versions.
In these cases, it is easier to allow a dirty state for faster
iteration. A parameter has been added to the recipes to allow them to
run whether or not git is clean.