Declare all dependencies in workspace
All dependencies and their versions are now centrally managed in the
workspace's `Cargo.toml`, and referenced in the crates as workspace
dependencies. This makes it easier to ensure we use a single,
consistent version of each crate and it provides a holistic overview of
all the dependencies used by clawless.