Reintroduce `commands` module
After some deliberation and experimentation, we have decided to bring
back the `commands` module. Especially in smaller CLIs, e.g. for local
development tasks, it is often overkill to separate the CLI and the
business logic into different crates. In such applications, it is
beneficial to have all commands scoped under a single module to clearly
separate them from the rest of the application. And for bigger
applications, we expect future features (e.g. configuration management)
to provide enough other content that the `commands` module does not feel
arbitrary.