Pass a `Context` to commands
The public interface for commands has been changed so that they accept
both command-line arguments as well as a context object. The context
provides information about the environment that a command runs in, and
access to shared resources and configuration settings.

A requirement that this change introduces is that commands MUST accept
those two arguments. This will be checked and reported using better
error messages in a future commit.