Sort imports when formatting Rust code (#137)
We are now using the nightly toolchain with the unstable group imports
option to format Rust code. This option creates three groups for `use`
statements: imports from the standard library, external crates, and
imports from the crate itself.

The reason why we do this is to ensure consistency across developers. By
letting rustfmt deal with imports, we reduce churn and keep pull request
reviews focused on the important changes.