Rename the `noop` attribute to `require_subcommand` (#144)
The `noop` attribute for the `#[command]` macro was neither well
documented nor well named. Both have been fixed. The attribute has been
renamed to `require_subcommand`, which better describes its purpose. If
the argument is added, the command function will never be executed and
instead passing in a subcommand is mandatory.